some warnings suppression gcc 4.4 'above array bound' missing in op3x3
git-svn-id: svn+ssh://gforge/svn/siconos/trunk@2326 bf01a9d7-0df3-0310-8ad4-b1c94d9ae2d5

+++ Numerics/src/netlib/odepack/test/funC.cpp
@@ -37,7 +37,7 @@ extern "C" void F77NAME(f1)(integer *sizeOfX, doublereal *time, doublereal *x, d
   }
   else
   {
-    printf("OscillatorPlugin:vectorField --- Bad size of x. %ld -- %p\n", *sizeOfX, sizeOfX);
+    printf("OscillatorPlugin:vectorField --- Bad size of x. %ld -- %p\n", (long int) *sizeOfX, sizeOfX);
   }
 
 }
